Remove orchestrator V1
